Fix some bad casting in error reporting
For larger-architectures, it's probably more advisable to use a long.
The types were also wrong, because casting has a higher precedence
order than the infix operator '+'.

The format string %Zu is the cleanest solution, but not defined in
many compilers and in C89, so compromise using this well-known cast
heuristic that works 'pretty well' on 'most platforms' and 'most
compilers'.

+++ xlogdump.c
@@ -314,12 +314,13 @@ restart:
 		record->xl_tot_len > SizeOfXLogRecord + record->xl_len +
 		XLR_MAX_BKP_BLOCKS * (sizeof(BkpBlock) + BLCKSZ))
 	{
-		printf("invalid record length(expected %u ~ %u, actual %d) at %X/%X\n",
-				(unsigned int)SizeOfXLogRecord + record->xl_len,
-				(unsigned int)SizeOfXLogRecord + record->xl_len +
-					XLR_MAX_BKP_BLOCKS * (sizeof(BkpBlock) + BLCKSZ),
-				record->xl_tot_len,
-			   curRecPtr.xlogid, curRecPtr.xrecoff);
+		printf(
+			"invalid record length(expected %lu ~ %lu, actual %d) at %X/%X\n",
+			(unsigned long) (SizeOfXLogRecord + record->xl_len),
+			(unsigned long) (SizeOfXLogRecord + record->xl_len +
+							 XLR_MAX_BKP_BLOCKS * (sizeof(BkpBlock) + BLCKSZ)),
+			record->xl_tot_len,
+			curRecPtr.xlogid, curRecPtr.xrecoff);
 		return false;
 	}
 	total_len = record->xl_tot_len;
